+++ src/replacement_policy/least_recently_used.h
@@ -3,20 +3,20 @@
 
 #include <unordered_map>
 
-#include "doubly_linked_list/doubly_linked_list.h"
+#include "../doubly_linked_list/doubly_linked_list.h"
 #include "replacement_policy.h"
 
 class LeastRecentlyUsed : public ReplacementPolicy {
 public:
     LeastRecentlyUsed();
 
-    void update(uint64_t index);
-    uint64_t get_replacement_index();
-    void remove(uint64_t index);
+    void update(uint32_t index);
+    uint32_t get_replacement_index();
+    void remove(uint32_t index);
 
 private:
-    DoublyLinkedList<uint64_t> dll_;
-    std::unordered_map<uint64_t, std::shared_ptr<DoublyLinkedListNode<uint64_t>>> map_;
+    DoublyLinkedList<uint32_t> dll_;
+    std::unordered_map<uint32_t, std::shared_ptr<DoublyLinkedListNode<uint32_t>>> map_;
     std::unordered_map<int, int> map2_;
 
     uint64_t size_;
